+++ src/app/service/generic.service.ts
@@ -1,7 +1,6 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { environment } from '../../environments/environment';
-import { UserDetailsModel } from '../shared/models/util.model';
 import { AuthService } from './auth.service';
 
 @Injectable({
