+++ src/app/service/notification.service.ts
@@ -2,12 +2,17 @@ import { Injectable } from '@angular/core';
 import { Router } from '@angular/router';
 import Swal from 'sweetalert2';
 import { AuthService } from './auth.service';
+import { GenericService } from './generic.service';
 
 @Injectable({
   providedIn: 'root',
 })
 export class NotificationService {
-  constructor(private authService: AuthService, private router: Router) {}
+  constructor(
+    private genericService: GenericService,
+    private authService: AuthService,
+    private router: Router
+  ) {}
 
   public error(errMsg: string, logOutRequired: boolean = false): void {
     Swal.fire({
@@ -17,7 +22,7 @@ export class NotificationService {
       confirmButtonText: 'OK',
     }).then((res) => {
       if (logOutRequired) {
-        this.authService.logout();
+        this.logout();
       }
     });
   }
@@ -33,6 +38,15 @@ export class NotificationService {
           replaceUrl: true,
         });
       }
+    });
+  }
+
+  public logout() {
+    this.genericService.logoutApi().subscribe((data: any) => {
+      sessionStorage.clear();
+      this.router.navigate(['/login'], {
+        replaceUrl: true,
+      });
     });
   }
 }
