+++ src/app/shared/header/header.component.html
@@ -8,7 +8,7 @@
             <path d="M11 6a3 3 0 1 1-6 0 3 3 0 0 1 6 0z"/>
             <path fill-rule="evenodd" d="M0 8a8 8 0 1 1 16 0A8 8 0 0 1 0 8zm8-7a7 7 0 0 0-5.468 11.37C3.242 11.226 4.805 10 8 10s4.757 1.225 5.468 2.37A7 7 0 0 0 8 1z"/>
           </svg>
-          <span [innerHTML]="userDetails.user_role === 'admin' ? 'Admin' : userDetails?.user_name"></span>
+          <span [innerHTML]="userDetails?.user_role === 'admin' ? 'Admin' : userDetails?.user_name"></span>
         </a>
         <a (click)="logout()" style="cursor: pointer;">
           <span>Log Out</span>
